requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
channels==4.0.0
Django==5.0
django-ninja==1.1.0
mysqlclient==2.2.1
pydantic==2.5.3
pydantic_core==2.14.6
pymemcache==4.0.0
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.9.0
